backend/requirements.txt
# Backend Dependencies for Parsey TOEIC Service

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# OCR
google-cloud-vision>=3.5.0

# LLM
openai>=1.6.0
google-generativeai>=0.3.0

# RAG & Vector DB
chromadb>=0.4.22
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10

# Utilities
aiofiles>=23.2.1
httpx>=0.25.0
